requirements.txt
aiogram==3.10.0
supabase==2.5.0
python-dotenv==1.0.0
loguru==0.7.2